requirements.txt
alabaster==0.7.12
asttokens
Babel==2.10.3
backcall 
black==22.6.0
brotlipy==0.7.0
certifi 
cffi 
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.1.0
cryptography 
cycler 
debugpy==1.6.2
decorator
dill==0.3.5.1
docutils==0.18.1
entrypoints==0.4
executing
flake8==4.0.1
fonttools==4.25.0
future==0.18.2
hyperopt==0.2.7
idna 
imagesize==1.4.1
ipykernel==6.15.1
ipython
jedi
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.11.1
kiwisolver 
markdown-it-py==2.1.0
MarkupSafe==2.1.1
mccabe==0.6.1
mdit-py-plugins==0.3.0
mdurl==0.1.1
multiprocess==0.70.13
munkres==1.1.4
mypy-extensions==0.4.3
myst-parser==0.18.0
nest-asyncio==1.5.5
networkx==2.8.5
numpy==1.23.1
packaging 
parso 
pathos==0.2.9
pathspec==0.9.0
pexpect
pickleshare
Pillow==9.2.0
platformdirs==2.5.2
pox==0.3.1
ppft==1.7.6.5
prompt-toolkit 
psutil==5.9.1
ptyprocess 
pure-eval
py4j==0.10.9.5
pycodestyle==2.8.0
pycparser
pyflakes==2.4.0
Pygments==2.12.0
pyOpenSSL
pyparsing==3.0.9
PySocks 
python-dateutil
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
requests 
scipy==1.8.1
six 
snowballstemmer==2.2.0
Sphinx==5.1.0
sphinx-rtd-theme 
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data
threadpoolctl==3.1.0
tomli==2.0.1
#torch==1.12.0
tornado 
tqdm==4.64.0
traitlets
typing_extensions
urllib3==1.26.11
wcwidth
